+++ feature/info/src/main/java/cx/aswin/boxlore/feature/info/EpisodeInfoScreen.kt
@@ -6,7 +6,6 @@ import androidx.compose.animation.core.Spring
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.spring
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -28,20 +27,16 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.rounded.ArrowBack
 import androidx.compose.material.icons.automirrored.rounded.KeyboardArrowRight
 import androidx.compose.material.icons.automirrored.rounded.Label
 import androidx.compose.material.icons.outlined.Share
-import androidx.compose.material.icons.rounded.AutoAwesome
 import androidx.compose.material.icons.rounded.CalendarToday
 import androidx.compose.material.icons.rounded.Label
 import androidx.compose.material.icons.rounded.Schedule
-import androidx.compose.material.icons.rounded.Subscriptions
 import androidx.compose.material.icons.rounded.Tag
 import androidx.compose.material.icons.rounded.Videocam
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -64,7 +59,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.blur
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.layout.ContentScale
@@ -76,7 +70,6 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.lerp
-import androidx.compose.ui.unit.sp
 import androidx.palette.graphics.Palette
 import coil.compose.AsyncImagePainter
 import coil.compose.rememberAsyncImagePainter
@@ -85,7 +78,6 @@ import cx.aswin.boxlore.core.designsystem.components.BoxLoreLoader
 import cx.aswin.boxlore.core.designsystem.components.ControlStyle
 import cx.aswin.boxlore.core.designsystem.components.OptimizedImage
 import cx.aswin.boxlore.core.designsystem.theme.expressiveClickable
-import cx.aswin.boxlore.core.designsystem.theme.m3Shimmer
 import kotlinx.coroutines.delay
 
 // Color extraction helper
@@ -748,340 +740,23 @@ fun EpisodeInfoScreen(
                     // Contextual "MORE LIKE THIS" RECOMMENDATIONS SECTION -> Card
                     if (state.similarEpisodesLoading || state.similarEpisodes.isNotEmpty()) {
                         item {
-                            androidx.compose.material3.OutlinedCard(
-                                modifier =
-                                    Modifier
-                                        .fillMaxWidth()
-                                        .padding(horizontal = 16.dp),
-                                colors =
-                                    androidx.compose.material3.CardDefaults.outlinedCardColors(
-                                        containerColor = MaterialTheme.colorScheme.surfaceContainerLow,
-                                    ),
-                                border = BorderStroke(0.5.dp, MaterialTheme.colorScheme.outlineVariant),
-                                shape = MaterialTheme.shapes.extraLarge,
-                            ) {
-                                Column(
-                                    modifier =
-                                        Modifier
-                                            .fillMaxWidth()
-                                            .padding(vertical = 16.dp),
-                                ) {
-                                    Row(
-                                        modifier =
-                                            Modifier
-                                                .fillMaxWidth()
-                                                .padding(horizontal = 20.dp)
-                                                .padding(bottom = 12.dp),
-                                        verticalAlignment = Alignment.CenterVertically,
-                                    ) {
-                                        Icon(
-                                            imageVector = Icons.Rounded.AutoAwesome,
-                                            contentDescription = null,
-                                            tint = MaterialTheme.colorScheme.primary,
-                                            modifier = Modifier.size(22.dp),
-                                        )
-                                        Spacer(modifier = Modifier.width(10.dp))
-                                        Text(
-                                            text = "More Like This",
-                                            style =
-                                                MaterialTheme.typography.titleMedium.copy(
-                                                    fontWeight = FontWeight.Bold,
-                                                    letterSpacing = (-0.1).sp,
-                                                ),
-                                            color = MaterialTheme.colorScheme.onSurface,
-                                            modifier = Modifier.weight(1f),
-                                            maxLines = 1,
-                                            overflow = TextOverflow.Ellipsis,
-                                        )
-                                    }
-
-                                    val similarListState = rememberLazyListState()
-                                    LazyRow(
-                                        state = similarListState,
-                                        horizontalArrangement = Arrangement.spacedBy(12.dp),
-                                        contentPadding = PaddingValues(horizontal = 16.dp),
-                                    ) {
-                                        if (state.similarEpisodesLoading) {
-                                            items(4) {
-                                                val baseColor = MaterialTheme.colorScheme.surfaceContainerHigh
-                                                val highlightColor = MaterialTheme.colorScheme.surfaceContainerHighest
-
-                                                Column(
-                                                    modifier = Modifier.width(120.dp),
-                                                    horizontalAlignment = Alignment.CenterHorizontally,
-                                                ) {
-                                                    Box(
-                                                        modifier =
-                                                            Modifier
-                                                                .size(120.dp)
-                                                                .clip(MaterialTheme.shapes.medium)
-                                                                .background(baseColor)
-                                                                .m3Shimmer(baseColor, highlightColor),
-                                                    )
-                                                    Spacer(modifier = Modifier.height(8.dp))
-                                                    Box(
-                                                        modifier =
-                                                            Modifier
-                                                                .fillMaxWidth()
-                                                                .height(14.dp)
-                                                                .clip(MaterialTheme.shapes.small)
-                                                                .background(baseColor)
-                                                                .m3Shimmer(baseColor, highlightColor),
-                                                    )
-                                                    Spacer(modifier = Modifier.height(4.dp))
-                                                    Box(
-                                                        modifier =
-                                                            Modifier
-                                                                .fillMaxWidth(0.7f)
-                                                                .height(14.dp)
-                                                                .clip(MaterialTheme.shapes.small)
-                                                                .background(baseColor)
-                                                                .m3Shimmer(baseColor, highlightColor),
-                                                    )
-                                                }
-                                            }
-                                        } else {
-                                            items(state.similarEpisodes) { episode ->
-                                                androidx.compose.material3.OutlinedCard(
-                                                    shape = RoundedCornerShape(16.dp),
-                                                    colors =
-                                                        androidx.compose.material3.CardDefaults.outlinedCardColors(
-                                                            containerColor = MaterialTheme.colorScheme.surfaceContainerLowest,
-                                                        ),
-                                                    border = BorderStroke(0.5.dp, MaterialTheme.colorScheme.outlineVariant),
-                                                    modifier =
-                                                        Modifier
-                                                            .width(140.dp)
-                                                            .expressiveClickable {
-                                                                onEpisodeClick(episode)
-                                                            },
-                                                ) {
-                                                    Column {
-                                                        OptimizedImage(
-                                                            url = episode.imageUrl?.ifEmpty { episode.podcastImageUrl },
-                                                            proxyWidth = 300,
-                                                            contentDescription = episode.title,
-                                                            modifier =
-                                                                Modifier
-                                                                    .size(140.dp)
-                                                                    .clip(RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp)),
-                                                            contentScale = ContentScale.Crop,
-                                                        )
-                                                        Column(
-                                                            modifier = Modifier.padding(10.dp),
-                                                            verticalArrangement = Arrangement.spacedBy(4.dp),
-                                                        ) {
-                                                            Text(
-                                                                text = episode.title,
-                                                                style =
-                                                                    MaterialTheme.typography.labelMedium.copy(
-                                                                        fontWeight = FontWeight.Bold,
-                                                                        lineHeight = 14.sp,
-                                                                    ),
-                                                                color = MaterialTheme.colorScheme.onSurface,
-                                                                minLines = 2,
-                                                                maxLines = 2,
-                                                                overflow = TextOverflow.Ellipsis,
-                                                            )
-                                                            val podTitle = episode.podcastTitle
-                                                            if (!podTitle.isNullOrEmpty()) {
-                                                                Text(
-                                                                    text = podTitle,
-                                                                    style =
-                                                                        MaterialTheme.typography.bodySmall.copy(
-                                                                            fontSize = 11.sp,
-                                                                            fontWeight = FontWeight.Medium,
-                                                                        ),
-                                                                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                                                    maxLines = 1,
-                                                                    overflow = TextOverflow.Ellipsis,
-                                                                )
-                                                            }
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
+                            cx.aswin.boxlore.feature.info.sections.EpisodeInfoMoreLikeThisCard(
+                                state = state,
+                                onEpisodeClick = onEpisodeClick,
+                            )
                         }
                     }
 
                     // UNIFIED "MORE FROM PODCAST" SECTION -> Card
                     item {
-                        androidx.compose.material3.OutlinedCard(
-                            modifier =
-                                Modifier
-                                    .fillMaxWidth()
-                                    .padding(horizontal = 16.dp),
-                            colors =
-                                androidx.compose.material3.CardDefaults.outlinedCardColors(
-                                    containerColor = MaterialTheme.colorScheme.surfaceContainerLow,
-                                ),
-                            border = BorderStroke(0.5.dp, MaterialTheme.colorScheme.outlineVariant),
-                            shape = MaterialTheme.shapes.extraLarge,
-                        ) {
-                            Column(
-                                modifier =
-                                    Modifier
-                                        .fillMaxWidth()
-                                        .padding(vertical = 16.dp),
-                            ) {
-                                Row(
-                                    modifier =
-                                        Modifier
-                                            .fillMaxWidth()
-                                            .expressiveClickable {
-                                                viewModel.onPodcastLinkClicked()
-                                                onPodcastClick(state.podcastId)
-                                            }.padding(horizontal = 20.dp)
-                                            .padding(bottom = 12.dp),
-                                    verticalAlignment = Alignment.CenterVertically,
-                                ) {
-                                    Icon(
-                                        imageVector = Icons.Rounded.Subscriptions,
-                                        contentDescription = null,
-                                        tint = MaterialTheme.colorScheme.primary,
-                                        modifier = Modifier.size(20.dp),
-                                    )
-                                    Spacer(modifier = Modifier.width(10.dp))
-                                    Text(
-                                        text = "More from ${state.podcastTitle}",
-                                        style =
-                                            MaterialTheme.typography.titleMedium.copy(
-                                                fontWeight = FontWeight.Bold,
-                                                letterSpacing = (-0.1).sp,
-                                            ),
-                                        color = MaterialTheme.colorScheme.onSurface,
-                                        modifier = Modifier.weight(1f),
-                                        maxLines = 1,
-                                        overflow = TextOverflow.Ellipsis,
-                                    )
-                                    Icon(
-                                        imageVector = Icons.AutoMirrored.Rounded.KeyboardArrowRight,
-                                        contentDescription = "Go to podcast",
-                                        tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                                    )
-                                }
-
-                                // Horizontal episodes row
-                                val relatedListState = rememberLazyListState()
-                                LaunchedEffect(relatedListState.isScrollInProgress) {
-                                    if (relatedListState.isScrollInProgress) {
-                                        viewModel.onRelatedEpisodesScrolled()
-                                    }
-                                }
-                                LazyRow(
-                                    state = relatedListState,
-                                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                                    contentPadding = PaddingValues(horizontal = 16.dp),
-                                ) {
-                                    if (state.relatedEpisodesLoading) {
-                                        // Skeleton loaders
-                                        items(4) {
-                                            val baseColor = MaterialTheme.colorScheme.surfaceContainerHigh
-                                            val highlightColor = MaterialTheme.colorScheme.surfaceContainerHighest
-
-                                            Column(
-                                                modifier = Modifier.width(120.dp),
-                                                horizontalAlignment = Alignment.CenterHorizontally,
-                                            ) {
-                                                // Skeleton artwork with shimmer
-                                                Box(
-                                                    modifier =
-                                                        Modifier
-                                                            .size(120.dp)
-                                                            .clip(MaterialTheme.shapes.medium)
-                                                            .background(baseColor)
-                                                            .m3Shimmer(baseColor, highlightColor),
-                                                )
-
-                                                Spacer(modifier = Modifier.height(8.dp))
-
-                                                // Skeleton text with shimmer
-                                                Box(
-                                                    modifier =
-                                                        Modifier
-                                                            .fillMaxWidth()
-                                                            .height(14.dp)
-                                                            .clip(MaterialTheme.shapes.small)
-                                                            .background(baseColor)
-                                                            .m3Shimmer(baseColor, highlightColor),
-                                                )
-                                                Spacer(modifier = Modifier.height(4.dp))
-                                                Box(
-                                                    modifier =
-                                                        Modifier
-                                                            .fillMaxWidth(0.7f)
-                                                            .height(14.dp)
-                                                            .clip(MaterialTheme.shapes.small)
-                                                            .background(baseColor)
-                                                            .m3Shimmer(baseColor, highlightColor),
-                                                )
-                                            }
-                                        }
-                                    } else if (state.relatedEpisodes.isNotEmpty()) {
-                                        items(state.relatedEpisodes) { episode ->
-                                            androidx.compose.material3.OutlinedCard(
-                                                shape = RoundedCornerShape(16.dp),
-                                                colors =
-                                                    androidx.compose.material3.CardDefaults.outlinedCardColors(
-                                                        containerColor = MaterialTheme.colorScheme.surfaceContainerLowest,
-                                                    ),
-                                                border = BorderStroke(0.5.dp, MaterialTheme.colorScheme.outlineVariant),
-                                                modifier =
-                                                    Modifier
-                                                        .width(140.dp)
-                                                        .expressiveClickable {
-                                                            viewModel.onRelatedEpisodeClicked()
-                                                            onEpisodeClick(episode)
-                                                        },
-                                            ) {
-                                                Column {
-                                                    // Episode Artwork
-                                                    OptimizedImage(
-                                                        url = episode.imageUrl?.ifEmpty { state.episode.podcastImageUrl },
-                                                        proxyWidth = 300, // 140dp thumbnails
-                                                        contentDescription = episode.title,
-                                                        modifier =
-                                                            Modifier
-                                                                .size(140.dp)
-                                                                .clip(RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp)),
-                                                        contentScale = ContentScale.Crop,
-                                                    )
-
-                                                    // Title in card footer - minLines for even sizing
-                                                    Text(
-                                                        text = episode.title,
-                                                        style =
-                                                            MaterialTheme.typography.labelMedium.copy(
-                                                                fontWeight = FontWeight.SemiBold,
-                                                                lineHeight = 14.sp,
-                                                            ),
-                                                        color = MaterialTheme.colorScheme.onSurface,
-                                                        minLines = 3,
-                                                        maxLines = 3,
-                                                        overflow = TextOverflow.Ellipsis,
-                                                        modifier = Modifier.padding(12.dp),
-                                                    )
-                                                }
-                                            }
-                                        }
-                                    } else {
-                                        // No episodes message
-                                        item {
-                                            Text(
-                                                text = "No other episodes available",
-                                                style = MaterialTheme.typography.bodyMedium,
-                                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                                modifier = Modifier.padding(vertical = 16.dp),
-                                            )
-                                        }
-                                    }
-                                }
-                            }
-                        }
+                        cx.aswin.boxlore.feature.info.sections.EpisodeInfoMoreFromPodcastCard(
+                            state = state,
+                            onPodcastClick = onPodcastClick,
+                            onEpisodeClick = onEpisodeClick,
+                            onPodcastLinkClicked = viewModel::onPodcastLinkClicked,
+                            onRelatedEpisodesScrolled = viewModel::onRelatedEpisodesScrolled,
+                            onRelatedEpisodeClicked = viewModel::onRelatedEpisodeClicked,
+                        )
                     }
                 }
             }
